Split tests into multiple test files
